Re-write description as 3-parts over 3 days
As written, I don't understand what this means:

> Python Fundamentals consists of two parts, both split into three sessions. Each session takes 2 hours.

So I've rewritten this to advertise in our newsletter in a way that follows how we describe other workshops.
Each day is its own Part X (1, 2, 3). So this is a 3-part workshop (not two parts, even if there is a two-part conceptual split).